chore(routes-schema): .int() для id/createdAt в response schemas
Завершение integer-validation серии: добавил .int() на response-side
для invitations.create.responses[201].id/createdAt и
invitations.accept.responses[201].company.id.

Defensive: server гарантированно возвращает integer (autoincrement
column), но если БД будет corrupted (manual ALTER, restore из бекапа
с ошибкой) или server отдаст float через bug — Zod на client'е
свалится с понятным ZodError, а не silently передаст float дальше
в state, где `.toString()` или арифметика дадут странные результаты.

Без изменения семантики happy-path. (typecheck зелёный, 1261 тестов
зелёные)

+++ shared/routes.ts
@@ -86,12 +86,12 @@ export const api = {
       }),
       responses: {
         201: z.object({
-          id: z.number(),
+          id: z.number().int(),
           token: z.string(),
           url: z.string(),
           position: z.string().nullable(),
           isAdmin: z.boolean(),
-          createdAt: z.number(),
+          createdAt: z.number().int(),
         }),
         400: errorSchemas.validation,
       },
@@ -125,7 +125,7 @@ export const api = {
       responses: {
         201: z.object({
           user: z.custom<typeof users.$inferSelect>(),
-          company: z.object({ id: z.number(), name: z.string() }),
+          company: z.object({ id: z.number().int(), name: z.string() }),
         }),
         400: errorSchemas.validation,
       },
